Clamp, widen, and saturate the remaining audit correctness lows
EventStream::retire_batch subtracted its batch size from the accountant's
outstanding count directly, where every other terminalization site routes
through record_left_outstanding, which floors at zero. Any state reporting
fewer outstanding than the batch holds -- a shared accountant whose other
flow terminalized first, or a reset between publish and retire -- wrapped
the unsigned counter. Measured at 18446744073709551614, which
inventory_tick_weighted then multiplies by. Diagnostics only, but the
corrupted values are the ones a health view reports.

save_world_archive let the constant's own type decide the width of
lattice_version, while LatticeType requires only convertibility to uint32.
A custom lattice -- a documented extension point -- declaring a uint64
version wrote a header four bytes too long against a fixed-width reader.
Verified by adding one: without the cast, inspect and load both reject the
file the writer had just reported Ok for.

The A* plane-gap and axis-detour fast paths now saturate. manhattan clamps
each term at uint32 max and the sums wrapped, asymmetric with
best_chunk_portal, which already saturated the identical expression. This
one carries no test: reaching the wrap needs a span beyond 2^32 tiles,
which cannot be allocated, so it is consistency rather than a live defect
and I would rather say so than write a test that cannot fail.

DeltaCollector sizes its coalesce table from the realized capacity instead
of the requested one. append_entity admits while size() != capacity() and
reserve(n) guarantees only capacity() >= n, so an implementation rounding
past 2 * entity_capacity could fill every slot and leave both unbounded
probe loops spinning. No shipped standard library does that, which is why
the invariant is now structural rather than resting on allocator
behaviour.

The two testable fixes are mutation-verified. Full suite and ASan are
green at 1117 tests each.

+++ include/tess/persistence/archive.h
@@ -805,8 +805,15 @@ auto save_world_archive(const World& world, std::vector<std::byte>& out)
   detail::append_unsigned_le(
       out, static_cast<std::uint32_t>(
                ShapeTraits<typename World::shape_type>::lattice_identity));
+  // Explicitly widened like the identity above. `append_unsigned_le` emits
+  // sizeof(UInt) bytes and `lattice_version` is `static constexpr auto`,
+  // while `LatticeType` only requires convertibility to uint32. A custom
+  // lattice declaring a uint64 version would otherwise write a 125-byte
+  // header against the fixed-width reader, producing a file that saves Ok
+  // and can never be loaded.
   detail::append_unsigned_le(
-      out, ShapeTraits<typename World::shape_type>::lattice_version);
+      out, static_cast<std::uint32_t>(
+               ShapeTraits<typename World::shape_type>::lattice_version));
   detail::append_unsigned_le(out, detail::world_archive_key_layout_version);
   detail::append_unsigned_le(out, Archive::id);
   detail::append_unsigned_le(out, Archive::version);
